streamlit/requirements.txt
streamlit==1.10.0 
requests==2.24.0
requests-toolbelt==0.9.1
scikit-learn==1.1.3
pandas==1.5.1
seaborn==0.12.2
nltk==3.8.1
numpy==1.23.4
matplotlib==3.7.0
Pillow==9.4.0
spacy==3.5.0
gensim==4.2.0
tensorflow==2.11.0



#pickle
#joblib
#webbrowser

#sys
#re
